+++ src/main/kotlin/blackjack/domain/model/card/Card.kt
@@ -1,4 +1,4 @@
-package blackjack.model.domain.card
+package blackjack.domain.model.card
 
 data class Card(
     val symbol: Shape,
