+++ src/main/kotlin/blackjack/domain/model/card/PlayingCard.kt
@@ -1,4 +1,4 @@
-package blackjack.model.domain.card
+package blackjack.domain.model.card
 
 class PlayingCard(private val deck: ArrayDeque<Card>) {
     fun spreadCard(): Card {
