+++ src/main/kotlin/blackjack/domain/model/participant/Dealer.kt
@@ -1,7 +1,7 @@
-package blackjack.model.domain.participant
+package blackjack.domain.model.participant
 
-import blackjack.model.domain.card.Card
-import blackjack.model.domain.card.Hand
+import blackjack.domain.model.card.Card
+import blackjack.domain.model.card.Hand
 
 class Dealer(override val name: String = DEALER_NAME) : Participants() {
     override val hand: Hand = Hand(mutableListOf())
