+++ src/test/kotlin/blackjack/domain/model/PlayingCardTest.kt
@@ -1,9 +1,9 @@
-package blackjack.model.domain
+package blackjack.domain.model
 
-import blackjack.model.domain.card.Card
-import blackjack.model.domain.card.CardNumber
-import blackjack.model.domain.card.PlayingCard
-import blackjack.model.domain.card.Shape
+import blackjack.domain.model.card.Card
+import blackjack.domain.model.card.CardNumber
+import blackjack.domain.model.card.PlayingCard
+import blackjack.domain.model.card.Shape
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
